add support for query facets
git-svn-id: svn://localhost/jahia/trunk@6036 69c369ed-ed09-0410-a4a1-e19a46c23626

+++ core/src/main/java/org/jahia/services/query/QueryResultWrapper.java
@@ -266,7 +266,7 @@ public class QueryResultWrapper implements QueryResult {
                 : null;
     }
 
-    public Map<String, Integer> getFacetQuery() {
+    public Map<String, Long> getFacetQuery() {
         return result instanceof JahiaMultiColumnQueryResult ? ((JahiaMultiColumnQueryResult) result)
                 .getFacetQuery()
                 : null;
